xsdk: Use only <cflags> and <linkflags> in BSP and application.
This avoids having to do a lot of processing in the generator to
handle every case possible.  It is better to let the user specify the
properties they desire and let the build fail naturally than try to
micromanage everything.

Note that there is likely more work to do here in case <cflags> and
<linkflags> are not set as is the case when features are supported
directly in the toolset.